+++ src/Application/Events/EventDispatcherInterface.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace Iriven\PhpFormGenerator\Application\Events;
+
+/** @api */
+interface EventDispatcherInterface
+{
+    public function dispatch(object $event): void;
+}