+++ src/Application/Events/FormRenderEvent.php
@@ -1,0 +1,19 @@
+<?php
+declare(strict_types=1);
+
+namespace Iriven\PhpFormGenerator\Application\Events;
+
+use Iriven\PhpFormGenerator\Domain\Form\Form;
+
+/** @api */
+final class FormRenderEvent
+{
+    /**
+     * @param array<string, mixed> $context
+     */
+    public function __construct(
+        public Form $form,
+        public array $context = [],
+    ) {
+    }
+}