+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 ## Versions
 
+- [V6.2.0](releases.d/RELEASE_NOTES_V6.2.0.md)
 - [V6.1.0](releases.d/RELEASE_NOTES_V6.1.0.md)
 - [V6.0.0](releases.d/RELEASE_NOTES_V6.0.0.md)
 - [V5.9.1](releases.d/RELEASE_NOTES_V5.9.1.md)